Break XrdHttp into a module and a utility library.
The structure of the extension handlers is such that they must link
against some of the XrdHttp code.  Previously, the only way to do this
was to link against the shared module itself.

With this, we create and publish a proper versioned utility library.
